test(#5071): extract relay delivery seam
The maintainability gate was red because session_relay_sink.rs had 1514 production LoC against its 1499 frozen baseline, and the inventory gate was red because gateway.rs reached 1065 production lines without giant-file registration.

Extract the delegated-terminal delivery commit into delivery_commit.rs and move the five orchestration tests and fake transport into delivery_orchestration_tests.rs. Keep the production transport and terminal receipt-to-done path behavior-preserving while making the delivery writer seam injectable only in tests.

session_relay_sink.rs now measures 1442 production LoC, 57 below its 1499 baseline, so no baseline bump is needed.
