+++ justfile
@@ -34,6 +34,8 @@ test-non-pg:
     cargo test --lib source_registry -- --skip _pg --skip pg_ --skip postgres
     cargo test --lib task_notification -- --skip _pg --skip pg_ --skip postgres
     cargo test --lib delivery_lease_key -- --skip _pg --skip pg_ --skip postgres
+    # #5071 T0: keep the delivery-writer/terminal-fold contract seam in a curated lane.
+    cargo test --lib services::discord::session_relay_sink::delivery_orchestration_tests -- --skip _pg --skip pg_ --skip postgres
     cargo test --lib services::discord::e2e_control::tests -- --skip _pg --skip pg_ --skip postgres
     cargo test --lib server::routes::e2e_control::tests -- --skip _pg --skip pg_ --skip postgres
     cargo test --lib formatting -- --skip _pg --skip pg_ --skip postgres
