+++ src/services/discord/session_relay_sink.rs
@@ -37,6 +37,7 @@ pub(in crate::services::discord) const PURE_SUBAGENT_ZERO_DELIVERY_PAYLOAD: &str
     "{\"type\":\"result\",\"subtype\":\"success\",\"result\":\"done\"}\n"
 );
 
+mod delivery_commit;
 mod delivery_frontier;
 mod delivery_outcome_classify;
 mod idle_jsonl;
@@ -512,9 +513,8 @@ pub(in crate::services::discord) struct SessionBoundDiscordRelaySink {
     by_session: Mutex<HashMap<String, SessionRelayParser>>,
     #[cfg(test)]
     lease_test_probe: Option<Arc<SinkLeaseTestProbe>>,
-    test_gateway: Option<Arc<dyn super::gateway::TurnGateway>>,
     #[cfg(test)]
-    test_delivery_outcomes: Option<Arc<Mutex<Vec<SessionRelayDeliveryOutcome>>>>,
+    test_gateway: Option<Arc<dyn super::gateway::TurnGateway>>,
 }
 
 impl SessionBoundDiscordRelaySink {
@@ -526,9 +526,8 @@ impl SessionBoundDiscordRelaySink {
             by_session: Mutex::new(HashMap::new()),
             #[cfg(test)]
             lease_test_probe: None,
-            test_gateway: None,
             #[cfg(test)]
-            test_delivery_outcomes: None,
+            test_gateway: None,
         }
     }
 
@@ -685,108 +684,14 @@ impl SessionBoundDiscordRelaySink {
         }
     }
 
-    fn advance_offset_for_confirmed_delegated_terminal(
-        &self,
-        shared: &super::SharedData,
-        provider: &ProviderKind,
-        channel_id: u64,
-        session_name: &str,
-        delivery: &SessionRelayDelivery,
-        inflight: Option<&super::inflight::InflightTurnState>,
-    ) -> bool {
-        let Some(end) = delivery.terminal_consumed_end.filter(|end| *end > 0) else {
-            return false;
-        };
-        // IDENTITY GATE: the frame's pinned turn identity must still match the
-        // channel's current inflight. A delayed frame from an already-replaced
-        // turn (or a cleared inflight) is ignored — never advances a wrong turn.
-        let Some(inflight) = inflight else {
-            tracing::debug!(
-                provider = provider.as_str(),
-                channel_id,
-                tmux_session = %session_name,
-                frame_user_msg_id = delivery.frame_turn_user_msg_id,
-                "session-bound sink: terminal frame carried a commit fence but inflight is gone; identity gate blocks advance"
-            );
-            return false;
-        };
-        // #3041 P1-3 (codex P1-3 issue 2 R4): STRICT `turn_start_offset` identity — a
-        // REQUIRED gate part with NO None fallback (two `user_msg_id == 0` turns in the same
-        // second collide on the weak `(user_msg_id, started_at)` pair). A fenced frame is
-        // GUARANTEED a real offset by the producer, so `None`/mismatch is a stale/wrong-turn
-        // frame → MUST NOT advance (the watcher's SendFull delivers — no black-hole).
-        let identity_matches = inflight.user_msg_id == delivery.frame_turn_user_msg_id
-            && inflight.started_at == delivery.frame_turn_started_at
-            && delivery.frame_turn_start_offset.is_some()
-            && inflight.turn_start_offset == delivery.frame_turn_start_offset;
-        if !identity_matches {
-            tracing::debug!(
-                provider = provider.as_str(),
-                channel_id,
-                tmux_session = %session_name,
-                frame_user_msg_id = delivery.frame_turn_user_msg_id,
-                inflight_user_msg_id = inflight.user_msg_id,
-                frame_turn_start_offset = delivery.frame_turn_start_offset,
-                inflight_turn_start_offset = inflight.turn_start_offset,
-                "session-bound sink: terminal frame identity != current inflight; identity gate blocks advance (delayed/wrong-turn frame)"
-            );
-            return false;
-        }
-        super::tmux::advance_watcher_confirmed_end(
-            shared,
-            provider,
-            ChannelId::new(channel_id),
-            session_name,
-            end,
-            "src/services/discord/session_relay_sink.rs:sink_confirmed_terminal_advance",
-        );
-        // #3976: stamp the durable per-row delivered marker ONLY here — past the
-        // identity gate, after the `confirmed_end_offset` watermark advance fired
-        // (so a refused/identity-mismatched advance, which returned above, never
-        // marks the row). The watermark is resettable and writes nothing else to
-        // the row, so without this durable marker a delivered-but-unmirrored row is
-        // indistinguishable from a never-delivered black-hole and orphan-reclaim
-        // would re-emit its tail on a watermark reset. The flock RMW re-gates the
-        // identity under the lock, so a turn replaced during the POST is never
-        // marked. Best-effort: a residual crash between the POST and this write
-        // reverts the row to orphan shape on reboot (same at-most-once residual the
-        // #3918 marker bounds) — acceptable and no worse than today.
-        super::inflight::mark_session_bound_relay_delivered_locked(
-            provider,
-            channel_id,
-            &super::inflight::InflightTurnIdentity::from_state(inflight),
-            session_name,
-        );
-        true
-    }
-
-    /// #3089 A2b: short-replace via the turn-output controller, behaviourally equal to legacy
-    /// `replace_long_message_raw_with_outcome` — SAME transport + `LeaseHolder::Sink` cell (one
-    /// acquire/commit/release, no double-acquire), #3151 heartbeat, #2757 `PreserveAlways`,
-    /// `CommitOnFallback`, identity-gated advance (FRESH post-POST reload): confirmed POST →
-    /// `Delivered`, ambiguous → `Err(Transient)` (I2); `Replace { Active }` → `post_send_finalize`
-    /// no-op. `gateway` seam (review-fix Medium-1): live = real gateway, test fakes it.
-    async fn deliver_short_replace_via_controller<G: super::gateway::TurnGateway + ?Sized>(
-        &self,
-        gateway: &G,
-        ctx: short_controller::SinkShortReplaceCtx<'_>,
-    ) -> Result<SessionRelayDeliveryOutcome, RelaySinkError> {
-        short_controller::deliver_short_replace_via_controller(gateway, ctx).await
-    }
-
     async fn deliver_response(
         &self,
         delivery: SessionRelayDelivery,
     ) -> Result<SessionRelayDeliveryOutcome, RelaySinkError> {
-        self.deliver_response_via_gateway(delivery, self.test_gateway.clone())
-            .await
-    }
-
-    async fn deliver_response_via_gateway(
-        &self,
-        delivery: SessionRelayDelivery,
-        injected_gateway: Option<Arc<dyn super::gateway::TurnGateway>>,
-    ) -> Result<SessionRelayDeliveryOutcome, RelaySinkError> {
+        #[cfg(test)]
+        let gateway: Option<&dyn super::gateway::TurnGateway> = self.test_gateway.as_deref();
+        #[cfg(not(test))]
+        let gateway: Option<&dyn super::gateway::TurnGateway> = None;
         let channel_id = delivery.channel_id;
         let provider = delivery.provider.clone();
         let inflight = super::inflight::load_inflight_state(&provider, channel_id);
@@ -883,13 +788,11 @@ impl SessionBoundDiscordRelaySink {
         // route from PlaceholderEdit to NewMessage. Keep that entire operation on
         // the outer lease so both the card transport and the final answer remain
         // guarded even though the route mutates inside `ensure_card_and_route`.
-        let has_injected_gateway = injected_gateway.is_some();
         let cutover_short_replace = delivery.task_notification_context.is_none()
             && cutover_range.is_some()
             && !relay_text.is_empty()
             && matches!(route, SessionBoundTerminalDeliveryRoute::PlaceholderEdit(_))
-            && !session_bound_should_send_new_chunks_for_placeholder(&relay_text)
-            && !has_injected_gateway;
+            && !session_bound_should_send_new_chunks_for_placeholder(&relay_text);
         let (lease_range, lease_fallback_start) = sink_delivery_lease_coordinate(&delivery);
         let sink_lease_key = delivery_lease_key_for_frame(
             channel,
@@ -963,25 +866,15 @@ impl SessionBoundDiscordRelaySink {
             }
         }
 
-        let http = shared.serenity_http_or_token_fallback();
-        let live_gateway;
-        let gateway: &dyn super::gateway::TurnGateway = match injected_gateway.as_deref() {
-            Some(gateway) => gateway,
-            None => {
-                let http = http.clone().ok_or_else(|| {
-                    RelaySinkError::Transient(format!(
-                        "discord http unavailable for provider {}",
-                        provider.as_str()
-                    ))
-                })?;
-                live_gateway = super::gateway::DiscordGateway::new(
-                    http,
-                    shared.clone(),
-                    provider.clone(),
-                    None,
-                );
-                &live_gateway
-            }
+        let http = if gateway.is_some() {
+            Arc::new(serenity::http::Http::new("test-gateway"))
+        } else {
+            shared.serenity_http_or_token_fallback().ok_or_else(|| {
+                RelaySinkError::Transient(format!(
+                    "discord http unavailable for provider {}",
+                    provider.as_str()
+                ))
+            })?
         };
         let (route, task_card_message_id, task_response_claim_outcome) =
             ensure_card_and_route(&self.health_registry, &shared, &delivery, route).await?;
@@ -1027,33 +920,54 @@ impl SessionBoundDiscordRelaySink {
 
         if let SessionBoundTerminalDeliveryRoute::PlaceholderEdit(msg_id) = route {
             if let Some((start, end)) = cutover_range.filter(|_| cutover_short_replace) {
-                return self
-                    .deliver_short_replace_via_controller(
-                        gateway,
-                        short_controller::SinkShortReplaceCtx {
-                            shared: &shared,
-                            provider: &provider,
-                            channel,
-                            channel_id,
-                            msg_id,
-                            relay_text: &relay_text,
-                            delivered_fingerprint_body: &raw_response_text,
-                            delivery: &delivery,
-                            sink_lease_key,
-                            sink_delivery_authority,
-                            trace: &trace,
-                            range: (start, end),
-                            delivered_total: &self.delivered_total,
-                        },
-                    )
-                    .await;
+                let live_gateway = super::gateway::DiscordGateway::new(
+                    http.clone(),
+                    shared.clone(),
+                    provider.clone(),
+                    None,
+                );
+                return short_controller::deliver_short_replace_via_controller(
+                    gateway.unwrap_or(&live_gateway),
+                    short_controller::SinkShortReplaceCtx {
+                        shared: &shared,
+                        provider: &provider,
+                        channel,
+                        channel_id,
+                        msg_id,
+                        relay_text: &relay_text,
+                        delivered_fingerprint_body: &raw_response_text,
+                        delivery: &delivery,
+                        sink_lease_key,
+                        sink_delivery_authority,
+                        trace: &trace,
+                        range: (start, end),
+                        delivered_total: &self.delivered_total,
+                    },
+                )
+                .await;
             }
             if session_bound_should_send_new_chunks_for_placeholder(&relay_text) {
-                let message_ids = gateway
-                    .send_long_message_with_rollback(channel, msg_id, &relay_text)
+                let message_ids = if let Some(gateway) = gateway {
+                    gateway
+                        .send_long_message_with_rollback(channel, msg_id, &relay_text)
+                        .await
+                        .map_err(RelaySinkError::Transient)?
+                } else {
+                    formatting::send_long_message_raw_with_rollback(
+                        &http,
+                        channel,
+                        msg_id,
+                        &relay_text,
+                        &shared,
+                    )
                     .await
-                    .map_err(RelaySinkError::Transient)?;
-                let _ = gateway.delete_message(channel, msg_id).await;
+                    .map_err(|error| RelaySinkError::Transient(error.to_string()))?
+                };
+                if let Some(gateway) = gateway {
+                    let _ = gateway.delete_message(channel, msg_id).await;
+                } else {
+                    let _ = super::http::delete_channel_message(&http, channel, msg_id).await;
+                }
                 self.delivered_total.fetch_add(1, Ordering::AcqRel);
                 tracing::info!(
                     provider = provider.as_str(),
@@ -1091,11 +1005,25 @@ impl SessionBoundDiscordRelaySink {
                 );
                 return Ok(SessionRelayDeliveryOutcome::from_proof(proof));
             }
-            match gateway
-                .replace_message_with_anchor(channel, msg_id, &relay_text)
+            let mut last_chunk_anchor = None;
+            let replace_outcome = if let Some(gateway) = gateway {
+                gateway
+                    .replace_message_with_outcome(channel, msg_id, &relay_text)
+                    .await
+            } else {
+                formatting::replace_long_message_raw_with_outcome(
+                    &http,
+                    channel,
+                    msg_id,
+                    &relay_text,
+                    &shared,
+                    &mut last_chunk_anchor,
+                )
                 .await
-            {
-                Ok((ReplaceLongMessageOutcome::EditedOriginal, last_chunk_anchor)) => {
+                .map_err(|error| error.to_string())
+            };
+            match replace_outcome {
+                Ok(ReplaceLongMessageOutcome::EditedOriginal) => {
                     self.delivered_total.fetch_add(1, Ordering::AcqRel);
                     tracing::info!(
                         provider = provider.as_str(),
@@ -1139,13 +1067,10 @@ impl SessionBoundDiscordRelaySink {
                     );
                     Ok(SessionRelayDeliveryOutcome::from_proof(proof))
                 }
-                Ok((
-                    ReplaceLongMessageOutcome::SentFallbackAfterEditFailure {
-                        edit_error,
-                        replacement_anchor,
-                    },
-                    _,
-                )) => {
+                Ok(ReplaceLongMessageOutcome::SentFallbackAfterEditFailure {
+                    edit_error,
+                    replacement_anchor,
+                }) => {
                     // #2757 (A0 #3089): never delete msg_id — it is the bridge's
                     // current_msg_id, possibly holding streamed content a transient edit
                     // failure would vacuum. The shared policy pins this preserve decision.
@@ -1189,7 +1114,7 @@ impl SessionBoundDiscordRelaySink {
                     );
                     Ok(SessionRelayDeliveryOutcome::from_proof(proof))
                 }
-                Ok((ReplaceLongMessageOutcome::PartialContinuationFailure { error, .. }, _)) => {
+                Ok(ReplaceLongMessageOutcome::PartialContinuationFailure { error, .. }) => {
                     Err(RelaySinkError::Transient(
                         super::replace_outcome_policy::strip_watcher_send_failure_class_marker(
                             &error,
@@ -1525,13 +1450,16 @@ fn delivery_lease_key_for_frame(
 }
 
 #[cfg(test)]
+mod delivery_orchestration_tests;
+
+#[cfg(test)]
 mod tests {
     use super::*;
     use crate::services::cluster::session_matcher::{MatchedChannel, expected_rollout_path_for};
     use crate::services::discord::inflight::{RelayOwnerKind, TurnSource};
     use crate::services::tui_prompt_dedupe::{ExternalInputRelayLease, ExternalInputRelayOwner};
 
-    fn matched(channel_id: &str) -> MatchedChannel {
+    pub(super) fn matched(channel_id: &str) -> MatchedChannel {
         let session = ProviderKind::Claude.build_tmux_session_name(channel_id);
         MatchedChannel {
             channel_id: channel_id.to_string(),
@@ -1603,7 +1531,7 @@ mod tests {
     }
 
     #[allow(clippy::too_many_arguments)]
-    fn terminal_frame_offset(
+    pub(super) fn terminal_frame_offset(
         binding: &MatchedChannel,
         payload: &str,
         sequence: u64,
@@ -2633,252 +2561,6 @@ mod tests {
     }
 
     #[tokio::test]
-    async fn fenced_terminal_without_parser_delivery_is_terminal_not_delivered() {
-        let binding = matched("44001");
-        let sink = SessionBoundDiscordRelaySink::new(Arc::new(HealthRegistry::new()));
-        let terminal = terminal_frame_offset(
-            &binding,
-            "{\"type\":\"result\",\"result\":\"\"}\n",
-            1,
-            256,
-            0,
-            "2026-08-03T00:00:00Z",
-            Some(64),
-        );
-
-        let outcome = sink
-            .deliver(&terminal)
-            .await
-            .expect("a fenced terminal without parser delivery is known");
-
-        assert_eq!(outcome, RelaySinkOutcome::TerminalNotDelivered);
-    }
-
-    #[tokio::test]
-    async fn relay_deliver_propagates_injected_transport_error() {
-        let temp = tempfile::tempdir().expect("temp runtime root");
-        let _root = crate::config::set_agentdesk_root_for_test(temp.path());
-        let channel_id = 44_002;
-        let binding = matched(&channel_id.to_string());
-        let session = &binding.expected_session_name;
-        let generation_path =
-            crate::services::tmux_common::session_temp_path(session, "generation");
-        std::fs::create_dir_all(
-            std::path::Path::new(&generation_path)
-                .parent()
-                .expect("generation parent"),
-        )
-        .expect("generation directory");
-        std::fs::write(&generation_path, b"transport-error").expect("generation marker");
-        let started_at = "2026-08-03T00:00:01Z";
-        let mut inflight =
-            inflight_with_identity_offset(channel_id, session, 700, started_at, Some(0));
-        inflight.set_relay_owner_kind(RelayOwnerKind::SessionBoundRelay);
-        inflight.current_msg_id = 88_002;
-        super::super::inflight::save_inflight_state(&inflight).expect("persist inflight");
-        let registry = Arc::new(HealthRegistry::new());
-        let shared = super::super::make_shared_data_for_tests();
-        registry
-            .register(ProviderKind::Claude.as_str().to_string(), shared)
-            .await;
-        let gateway = Arc::new(RelayContractFakeGateway::failing("fake transport failure"));
-        let mut sink = SessionBoundDiscordRelaySink::new(registry);
-        sink.test_gateway = Some(gateway.clone());
-        let payload = concat!(
-            "{\"type\":\"assistant\",\"message\":{\"content\":[{\"type\":\"text\",\"text\":\"answer\"}]}}\n",
-            "{\"type\":\"result\",\"result\":\"answer\"}\n"
-        );
-        let terminal = terminal_frame_offset(&binding, payload, 1, 256, 700, started_at, Some(0));
-
-        let error = sink
-            .deliver(&terminal)
-            .await
-            .expect_err("transport failure must escape RelaySink::deliver");
-
-        assert!(matches!(error, RelaySinkError::Transient(_)), "{error:?}");
-        assert_eq!(gateway.replace_calls.load(Ordering::Acquire), 1);
-        super::super::inflight::clear_inflight_state(&ProviderKind::Claude, channel_id);
-    }
-
-    #[tokio::test]
-    async fn relay_deliver_preserves_tail_anchor_and_observes_persisted_proof() {
-        let temp = tempfile::tempdir().expect("temp runtime root");
-        let _root = crate::config::set_agentdesk_root_for_test(temp.path());
-        let channel_id = 44_003;
-        let binding = matched(&channel_id.to_string());
-        let session = &binding.expected_session_name;
-        let generation_path =
-            crate::services::tmux_common::session_temp_path(session, "generation");
-        std::fs::create_dir_all(
-            std::path::Path::new(&generation_path)
-                .parent()
-                .expect("generation parent"),
-        )
-        .expect("generation directory");
-        std::fs::write(&generation_path, b"persisted-proof").expect("generation marker");
-        let generation = dr::current_generation_mtime_ns(session);
-        let started_at = "2026-08-03T00:00:02Z";
-        let mut inflight =
-            inflight_with_identity_offset(channel_id, session, 701, started_at, Some(0));
-        inflight.set_relay_owner_kind(RelayOwnerKind::SessionBoundRelay);
-        inflight.current_msg_id = 88_003;
-        super::super::inflight::save_inflight_state(&inflight).expect("persist inflight");
-        let registry = Arc::new(HealthRegistry::new());
-        let shared = super::super::make_shared_data_for_tests();
-        registry
-            .register(ProviderKind::Claude.as_str().to_string(), shared.clone())
-            .await;
-        let gateway = Arc::new(RelayContractFakeGateway::edited(Some((
-            99_003,
-            "tail chunk",
-        ))));
-        let outcomes = Arc::new(Mutex::new(Vec::new()));
-        let mut sink = SessionBoundDiscordRelaySink::new(registry);
-        sink.test_gateway = Some(gateway.clone());
-        sink.test_delivery_outcomes = Some(outcomes.clone());
-        let payload = concat!(
-            "{\"type\":\"assistant\",\"message\":{\"content\":[{\"type\":\"text\",\"text\":\"answer\"}]}}\n",
-            "{\"type\":\"result\",\"result\":\"answer\"}\n"
-        );
-        let mut terminal =
-            terminal_frame_offset(&binding, payload, 1, 256, 701, started_at, Some(0));
-        terminal.relay_generation_mtime_ns = Some(generation);
-
-        let outcome = sink.deliver(&terminal).await.expect("persisted delivery");
-
-        assert_eq!(outcome, RelaySinkOutcome::TerminalDelivered);
-        assert_eq!(
-            outcomes.lock().expect("outcome probe").as_slice(),
-            &[SessionRelayDeliveryOutcome::Delivered]
-        );
-        let record = dr::read_record(&ProviderKind::Claude, channel_id).expect("delivery record");
-        assert_eq!(
-            record.delivered_frontier.expect("frontier").panel_msg_id,
-            Some(99_003),
-            "legacy replace must retain the formatter tail anchor"
-        );
-        assert_eq!(gateway.replace_calls.load(Ordering::Acquire), 1);
-        super::super::inflight::clear_inflight_state(&ProviderKind::Claude, channel_id);
-    }
-
-    #[tokio::test]
-    async fn relay_deliver_observes_landed_stale_proof() {
-        let temp = tempfile::tempdir().expect("temp runtime root");
-        let _root = crate::config::set_agentdesk_root_for_test(temp.path());
-        let channel_id = 44_004;
-        let binding = matched(&channel_id.to_string());
-        let session = &binding.expected_session_name;
-        let generation_path =
-            crate::services::tmux_common::session_temp_path(session, "generation");
-        std::fs::create_dir_all(
-            std::path::Path::new(&generation_path)
-                .parent()
-                .expect("generation parent"),
-        )
-        .expect("generation directory");
-        std::fs::write(&generation_path, b"landed-stale").expect("generation marker");
-        let generation = dr::current_generation_mtime_ns(session);
-        let started_at = "2026-08-03T00:00:03Z";
-        let mut inflight =
-            inflight_with_identity_offset(channel_id, session, 702, started_at, Some(0));
-        inflight.set_relay_owner_kind(RelayOwnerKind::SessionBoundRelay);
-        inflight.current_msg_id = 88_004;
-        super::super::inflight::save_inflight_state(&inflight).expect("persist inflight");
-        let registry = Arc::new(HealthRegistry::new());
-        let shared = super::super::make_shared_data_for_tests();
-        registry
-            .register(ProviderKind::Claude.as_str().to_string(), shared.clone())
-            .await;
-        let gateway = Arc::new(RelayContractFakeGateway {
-            on_transport: Some(Arc::new(move || {
-                super::super::inflight::clear_inflight_state(&ProviderKind::Claude, channel_id);
-            })),
-            ..RelayContractFakeGateway::edited(None)
-        });
-        let outcomes = Arc::new(Mutex::new(Vec::new()));
-        let mut sink = SessionBoundDiscordRelaySink::new(registry);
-        sink.test_gateway = Some(gateway);
-        sink.test_delivery_outcomes = Some(outcomes.clone());
-        let payload = concat!(
-            "{\"type\":\"assistant\",\"message\":{\"content\":[{\"type\":\"text\",\"text\":\"answer\"}]}}\n",
-            "{\"type\":\"result\",\"result\":\"answer\"}\n"
-        );
-        let mut terminal =
-            terminal_frame_offset(&binding, payload, 1, 256, 702, started_at, Some(0));
-        terminal.relay_generation_mtime_ns = Some(generation);
-
-        let outcome = sink
-            .deliver(&terminal)
-            .await
-            .expect("landed stale delivery");
-
-        assert_eq!(outcome, RelaySinkOutcome::TerminalDelivered);
-        assert_eq!(
-            outcomes.lock().expect("outcome probe").as_slice(),
-            &[SessionRelayDeliveryOutcome::LandedStale]
-        );
-        super::super::inflight::clear_inflight_state(&ProviderKind::Claude, channel_id);
-    }
-
-    #[tokio::test]
-    async fn relay_deliver_observes_landed_unrecorded_proof() {
-        let temp = tempfile::tempdir().expect("temp runtime root");
-        let _root = crate::config::set_agentdesk_root_for_test(temp.path());
-        let channel_id = 44_005;
-        let binding = matched(&channel_id.to_string());
-        let session = &binding.expected_session_name;
-        let generation_path =
-            crate::services::tmux_common::session_temp_path(session, "generation");
-        std::fs::create_dir_all(
-            std::path::Path::new(&generation_path)
-                .parent()
-                .expect("generation parent"),
-        )
-        .expect("generation directory");
-        std::fs::write(&generation_path, b"landed-unrecorded").expect("generation marker");
-        let generation = dr::current_generation_mtime_ns(session);
-        let started_at = "2026-08-03T00:00:04Z";
-        let mut inflight =
-            inflight_with_identity_offset(channel_id, session, 703, started_at, Some(0));
-        inflight.set_relay_owner_kind(RelayOwnerKind::SessionBoundRelay);
-        inflight.current_msg_id = 88_005;
-        super::super::inflight::save_inflight_state(&inflight).expect("persist inflight");
-        let runtime = temp.path().join("runtime");
-        std::fs::create_dir_all(&runtime).expect("runtime directory");
-        std::fs::write(runtime.join("discord_delivery_records"), b"not a directory")
-            .expect("block delivery record directory");
-        let registry = Arc::new(HealthRegistry::new());
-        let shared = super::super::make_shared_data_for_tests();
-        registry
-            .register(ProviderKind::Claude.as_str().to_string(), shared)
-            .await;
-        let gateway = Arc::new(RelayContractFakeGateway::edited(None));
-        let outcomes = Arc::new(Mutex::new(Vec::new()));
-        let mut sink = SessionBoundDiscordRelaySink::new(registry);
-        sink.test_gateway = Some(gateway);
-        sink.test_delivery_outcomes = Some(outcomes.clone());
-        let payload = concat!(
-            "{\"type\":\"assistant\",\"message\":{\"content\":[{\"type\":\"text\",\"text\":\"answer\"}]}}\n",
-            "{\"type\":\"result\",\"result\":\"answer\"}\n"
-        );
-        let mut terminal =
-            terminal_frame_offset(&binding, payload, 1, 256, 703, started_at, Some(0));
-        terminal.relay_generation_mtime_ns = Some(generation);
-
-        let outcome = sink
-            .deliver(&terminal)
-            .await
-            .expect("landed unrecorded delivery");
-
-        assert_eq!(outcome, RelaySinkOutcome::TerminalDelivered);
-        assert_eq!(
-            outcomes.lock().expect("outcome probe").as_slice(),
-            &[SessionRelayDeliveryOutcome::LandedUnrecorded]
-        );
-        super::super::inflight::clear_inflight_state(&ProviderKind::Claude, channel_id);
-    }
-
-    #[tokio::test]
     async fn inflightless_ranged_frame_acquires_sink_lease_before_terminal_delivery_4277() {
         let channel_id = 4_277_003;
         let binding = matched(&channel_id.to_string());
@@ -3414,154 +3096,6 @@ mod tests {
         }
     }
 
-    struct RelayContractFakeGateway {
-        replace_outcome: ReplaceLongMessageOutcome,
-        replace_anchor: Option<super::super::formatting::ReplaceLastChunkAnchor>,
-        transport_error: Option<String>,
-        sent_message_id: MessageId,
-        replace_calls: AtomicU64,
-        send_calls: AtomicU64,
-        on_transport: Option<Arc<dyn Fn() + Send + Sync>>,
-    }
-
-    impl RelayContractFakeGateway {
-        fn edited(anchor: Option<(u64, &str)>) -> Self {
-            Self {
-                replace_outcome: ReplaceLongMessageOutcome::EditedOriginal,
-                replace_anchor: anchor.map(|(msg_id, text)| {
-                    super::super::formatting::ReplaceLastChunkAnchor {
-                        msg_id,
-                        text: text.to_string(),
-                    }
-                }),
-                transport_error: None,
-                sent_message_id: MessageId::new(91_001),
-                replace_calls: AtomicU64::new(0),
-                send_calls: AtomicU64::new(0),
-                on_transport: None,
-            }
-        }
-
-        fn failing(message: &str) -> Self {
-            let mut gateway = Self::edited(None);
-            gateway.transport_error = Some(message.to_string());
-            gateway
-        }
-    }
-
-    impl super::super::gateway::TurnGateway for RelayContractFakeGateway {
-        fn send_message<'a>(
-            &'a self,
-            _channel_id: ChannelId,
-            _content: &'a str,
-        ) -> super::super::gateway::GatewayFuture<'a, Result<MessageId, String>> {
-            Box::pin(async move {
-                self.send_calls.fetch_add(1, Ordering::AcqRel);
-                if let Some(on_transport) = &self.on_transport {
-                    on_transport();
-                }
-                match &self.transport_error {
-                    Some(error) => Err(error.clone()),
-                    None => Ok(self.sent_message_id),
-                }
-            })
-        }
-
-        fn edit_message<'a>(
-            &'a self,
-            _channel_id: ChannelId,
-            _message_id: MessageId,
-            _content: &'a str,
-        ) -> super::super::gateway::GatewayFuture<'a, Result<(), String>> {
-            panic!("relay contract fake does not use edit_message")
-        }
-
-        fn replace_message_with_outcome<'a>(
-            &'a self,
-            _channel_id: ChannelId,
-            _message_id: MessageId,
-            _content: &'a str,
-        ) -> super::super::gateway::GatewayFuture<'a, Result<ReplaceLongMessageOutcome, String>>
-        {
-            Box::pin(async move {
-                self.replace_calls.fetch_add(1, Ordering::AcqRel);
-                if let Some(on_transport) = &self.on_transport {
-                    on_transport();
-                }
-                match &self.transport_error {
-                    Some(error) => Err(error.clone()),
-                    None => Ok(self.replace_outcome.clone()),
-                }
-            })
-        }
-
-        fn replace_message_with_anchor<'a>(
-            &'a self,
-            _channel_id: ChannelId,
-            _message_id: MessageId,
-            _content: &'a str,
-        ) -> super::super::gateway::GatewayFuture<
-            'a,
-            Result<
-                (
-                    ReplaceLongMessageOutcome,
-                    Option<super::super::formatting::ReplaceLastChunkAnchor>,
-                ),
-                String,
-            >,
-        > {
-            Box::pin(async move {
-                self.replace_calls.fetch_add(1, Ordering::AcqRel);
-                if let Some(on_transport) = &self.on_transport {
-                    on_transport();
-                }
-                match &self.transport_error {
-                    Some(error) => Err(error.clone()),
-                    None => Ok((self.replace_outcome.clone(), self.replace_anchor.clone())),
-                }
-            })
-        }
-
-        fn schedule_retry_with_history<'a>(
-            &'a self,
-            _channel_id: ChannelId,
-            _user_message_id: MessageId,
-            _user_text: &'a str,
-        ) -> super::super::gateway::GatewayFuture<'a, ()> {
-            panic!("relay contract fake does not schedule retries")
-        }
-
-        fn dispatch_queued_turn<'a>(
-            &'a self,
-            _channel_id: ChannelId,
-            _intervention: &'a super::super::Intervention,
-            _output_path: &'a str,
-            _skip_hook: bool,
-            _dispatch_lease: Option<Arc<crate::services::turn_orchestrator::DispatchLease>>,
-        ) -> super::super::gateway::GatewayFuture<'a, Result<(), String>> {
-            panic!("relay contract fake does not dispatch turns")
-        }
-
-        fn validate_live_routing<'a>(
-            &'a self,
-            _channel_id: ChannelId,
-        ) -> super::super::gateway::GatewayFuture<'a, Result<(), String>> {
-            panic!("relay contract fake does not validate routing")
-        }
-
-        fn requester_mention(&self) -> Option<String> {
-            None
-        }
-
-        fn can_chain_locally(&self) -> bool {
-            false
-        }
-
-        fn bot_owner_provider(&self) -> Option<ProviderKind> {
-            None
-        }
-    }
-
     struct NoopHeartbeatGuard;
     impl toc::PostHeartbeatGuard for NoopHeartbeatGuard {}
     struct NoopHeartbeat;
@@ -3780,7 +3314,7 @@ mod tests {
             (start, end),
         );
         let outcome = rt
-            .block_on(sink.deliver_short_replace_via_controller(
+            .block_on(short_controller::deliver_short_replace_via_controller(
                 &gateway,
                 short_controller::SinkShortReplaceCtx {
                     shared: &shared,
@@ -3849,7 +3383,7 @@ mod tests {
             (start, end),
         );
         let outcome2 = rt
-            .block_on(sink2.deliver_short_replace_via_controller(
+            .block_on(short_controller::deliver_short_replace_via_controller(
                 &gateway2,
                 short_controller::SinkShortReplaceCtx {
                     shared: &shared2,
@@ -3986,7 +3520,7 @@ mod tests {
         };
         let sink = SessionBoundDiscordRelaySink::new(Arc::new(HealthRegistry::new()));
         let outcome = rt
-            .block_on(sink.deliver_short_replace_via_controller(
+            .block_on(short_controller::deliver_short_replace_via_controller(
                 &gateway,
                 short_controller::SinkShortReplaceCtx {
                     shared: &shared,
@@ -4213,7 +3747,7 @@ mod tests {
         );
         assert!(
             module_src.contains(
-                "relay_text: &relay_text,\n                            delivered_fingerprint_body: &raw_response_text,"
+                "relay_text: &relay_text,\n                        delivered_fingerprint_body: &raw_response_text,"
             ),
             "session sink production cut-over call must thread the raw pre-format body into the fingerprint slot"
         );
@@ -4265,7 +3799,7 @@ mod tests {
         );
 
         let outcome = rt
-            .block_on(sink.deliver_short_replace_via_controller(
+            .block_on(short_controller::deliver_short_replace_via_controller(
                 &gateway,
                 short_controller::SinkShortReplaceCtx {
                     shared: &shared,
@@ -4439,7 +3973,7 @@ mod tests {
         inflight_with_identity_offset(channel_id, session_name, user_msg_id, started_at, None)
     }
 
-    fn inflight_with_identity_offset(
+    pub(super) fn inflight_with_identity_offset(
         channel_id: u64,
         session_name: &str,
         user_msg_id: u64,
